LPD-33733 Extract to a base test class and reuse
