+++ modules/apps/site-navigation/site-navigation-test/src/testIntegration/java/com/liferay/site/navigation/menu/web/internal/exportimport/test/SiteNavigationMenuExportImportTest.java
@@ -6,14 +6,10 @@
 package com.liferay.site.navigation.menu.web.internal.exportimport.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.dynamic.data.mapping.model.DDMTemplate;
-import com.liferay.dynamic.data.mapping.service.DDMTemplateLocalService;
-import com.liferay.dynamic.data.mapping.test.util.DDMTemplateTestUtil;
 import com.liferay.exportimport.kernel.configuration.ExportImportConfigurationParameterMapFactoryUtil;
 import com.liferay.exportimport.kernel.lar.PortletDataHandlerKeys;
 import com.liferay.exportimport.kernel.staging.MergeLayoutPrototypesThreadLocal;
 import com.liferay.exportimport.kernel.staging.StagingUtil;
-import com.liferay.exportimport.test.util.lar.BasePortletExportImportTestCase;
 import com.liferay.layout.test.util.LayoutTestUtil;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.model.Group;
@@ -21,25 +17,18 @@ import com.liferay.portal.kernel.model.Layout;
 import com.liferay.portal.kernel.model.LayoutSet;
 import com.liferay.portal.kernel.model.LayoutSetPrototype;
 import com.liferay.portal.kernel.portlet.PortletIdCodec;
-import com.liferay.portal.kernel.service.GroupLocalService;
 import com.liferay.portal.kernel.service.LayoutLocalService;
 import com.liferay.portal.kernel.service.PortletPreferencesLocalService;
-import com.liferay.portal.kernel.servlet.taglib.ui.LanguageEntry;
-import com.liferay.portal.kernel.template.TemplateConstants;
-import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.DeleteAfterTestRun;
 import com.liferay.portal.kernel.test.util.GroupTestUtil;
 import com.liferay.portal.kernel.test.util.RandomTestUtil;
 import com.liferay.portal.kernel.test.util.TestPropsValues;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.HashMapBuilder;
-import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.PortletKeys;
 import com.liferay.portal.kernel.util.UnicodeProperties;
 import com.liferay.portal.test.rule.Inject;
-import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
-import com.liferay.portal.test.rule.PermissionCheckerMethodTestRule;
-import com.liferay.portlet.display.template.PortletDisplayTemplate;
+import com.liferay.portlet.display.template.test.util.BaseExportImportTestCase;
 import com.liferay.site.navigation.constants.SiteNavigationMenuPortletKeys;
 import com.liferay.site.navigation.model.SiteNavigationMenu;
 import com.liferay.site.navigation.model.SiteNavigationMenuItem;
@@ -54,8 +43,6 @@ import javax.portlet.PortletPreferences;
 
 import org.junit.Assert;
 import org.junit.Before;
-import org.junit.ClassRule;
-import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
@@ -64,14 +51,7 @@ import org.junit.runner.RunWith;
  */
 @RunWith(Arquillian.class)
 public class SiteNavigationMenuExportImportTest
-	extends BasePortletExportImportTestCase {
-
-	@ClassRule
-	@Rule
-	public static final AggregateTestRule aggregateTestRule =
-		new AggregateTestRule(
-			new LiferayIntegrationTestRule(),
-			PermissionCheckerMethodTestRule.INSTANCE);
+	extends BaseExportImportTestCase {
 
 	@Override
 	public String getPortletId() throws Exception {
@@ -236,69 +216,6 @@ public class SiteNavigationMenuExportImportTest
 					Sites.MERGE_FAIL_COUNT)));
 	}
 
-	@Override
-	protected void testExportImportDisplayStyle(
-			long displayStyleGroupId, String scopeType)
-		throws Exception {
-
-		Group displayStyleGroup = _groupLocalService.getGroup(
-			displayStyleGroupId);
-
-		long classNameId = _portal.getClassNameId(
-			LanguageEntry.class.getName());
-
-		DDMTemplate ddmTemplate = DDMTemplateTestUtil.addTemplate(
-			displayStyleGroup.getGroupId(), classNameId, 0,
-			_portal.getClassNameId(PortletDisplayTemplate.class.getName()),
-			TemplateConstants.LANG_TYPE_FTL, RandomTestUtil.randomString(),
-			_portal.getSiteDefaultLocale(displayStyleGroup));
-
-		PortletPreferences portletPreferences = getImportedPortletPreferences(
-			HashMapBuilder.put(
-				"displayStyle",
-				new String[] {
-					PortletDisplayTemplate.DISPLAY_STYLE_PREFIX +
-						ddmTemplate.getTemplateKey()
-				}
-			).put(
-				"displayStyleGroupExternalReferenceCode",
-				() -> {
-					if (displayStyleGroup.getGroupId() == layout.getGroupId()) {
-						return null;
-					}
-
-					return new String[] {
-						displayStyleGroup.getExternalReferenceCode()
-					};
-				}
-			).build());
-
-		Assert.assertEquals(
-			PortletDisplayTemplate.DISPLAY_STYLE_PREFIX +
-				ddmTemplate.getTemplateKey(),
-			portletPreferences.getValue("displayStyle", null));
-
-		DDMTemplate importedDDMTemplate =
-			_ddmTemplateLocalService.fetchTemplate(
-				layout.getGroupId(), classNameId, ddmTemplate.getTemplateKey());
-
-		String importedDisplayStyleGroupExternalReferenceCode =
-			portletPreferences.getValue(
-				"displayStyleGroupExternalReferenceCode", null);
-
-		if (displayStyleGroup.getGroupId() != layout.getGroupId()) {
-			Assert.assertNull(importedDDMTemplate);
-
-			Assert.assertEquals(
-				displayStyleGroup.getExternalReferenceCode(),
-				importedDisplayStyleGroupExternalReferenceCode);
-		}
-		else {
-			Assert.assertNotNull(importedDDMTemplate);
-			Assert.assertNull(importedDisplayStyleGroupExternalReferenceCode);
-		}
-	}
-
 	private void _publishLayouts() throws Exception {
 		Map<String, String[]> parameterMap =
 			ExportImportConfigurationParameterMapFactoryUtil.
@@ -316,12 +233,6 @@ public class SiteNavigationMenuExportImportTest
 			_liveGroup.getGroupId(), false, parameterMap);
 	}
 
-	@Inject
-	private DDMTemplateLocalService _ddmTemplateLocalService;
-
-	@Inject
-	private GroupLocalService _groupLocalService;
-
 	private Layout _layout;
 
 	@Inject
@@ -329,9 +240,6 @@ public class SiteNavigationMenuExportImportTest
 
 	@DeleteAfterTestRun
 	private Group _liveGroup;
-
-	@Inject
-	private Portal _portal;
 
 	@Inject
 	private PortletPreferencesLocalService _portletPreferencesLocalService;
